requirements.txt
streamlit>=1.28.0
torch>=2.0.0
transformers>=4.50.0
Pillow>=9.0.0
pandas>=1.5.0